Передает данные в компонент/ы карточек предложений итерация 2. Создает компонент Перечня карточек предложений, добавляет функционал сохранения в свойство данного компонента id карточки на которую было произведено наведение.
